% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GamedayPredictions.R
\name{GamedayPredictions}
\alias{GamedayPredictions}
\title{GamedayPredictions}
\usage{
GamedayPredictions(model, test)
}
\arguments{
\item{model}{the fit glm model used to make predictions}

\item{test}{the test set to make predictions on}
}
\description{
GamedayPredictions
}
