% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model}
\alias{Model}
\title{Model}
\usage{
Model(train)
}
\arguments{
\item{train}{The dataset you'd like to train the model on (should come from MerlinXII::nba)}
}
\description{
This function fits a model to the most recent NBA data from 2018 to the last full gameday. Write the line strWL <- MerlinXII::FitModel and the fitted glm will be assigned to an object called strWL
}
