% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_History.R
\name{Model_History}
\alias{Model_History}
\title{Model_History}
\usage{
Model_History(data, form, n)
}
\arguments{
\item{data}{the preprocessed dataset to calculate performance off}

\item{form}{the referenced formula from a glm model to fit when making predictions}

\item{n}{the number of days back to be calculated}
}
\description{
Model_History
}
