% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA_current}
\alias{NBA_current}
\title{NBA_current}
\format{An object of class \code{data.frame} with 1376 rows and 33 columns.}
\usage{
NBA_current
}
\description{
Dataset of ONLY complete 2019-2020 NBA season (current NBA season).
}
\keyword{datasets}
