% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA_total}
\alias{NBA_total}
\title{NBA_total}
\format{An object of class \code{data.frame} with 3728 rows and 33 columns.}
\usage{
NBA_total
}
\description{
Dataset of all NBA game from both 2018-2019 NBA season and all complete games of 2019-2020 NBA season.
}
\keyword{datasets}
